requirements.txt
black--21.9b0 
pylint--2.11.1 
numpy--1.21.2
matplotlib--3.4.3 
python-osc-1.8.0
pywavelets-1.1.0
scikit-learn-1.0.1
pandas-1.3.4